requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-SocketIO==5.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Naked==0.1.32
numpy==2.1.3
pandas==2.2.3
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-engineio==4.11.1
python-socketio==5.12.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
simple-websocket==1.1.0
six==1.16.0
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
